add scss and move to folders
