+++ src/@batch-flask/ui/advanced-filter/advanced-filter.module.ts
@@ -5,8 +5,8 @@ import { MaterialModule } from "@batch-flask/core";
 import { FormModule } from "@batch-flask/ui/form";
 import { AdvancedFilterComponent } from "./advanced-filter.component";
 import { AdvancedFilterListComponent } from "./list-filter-control/list-filter-control";
-import { AdvancedFilterResultPickerComponent } from "./result-picker-control";
-import { AdvancedFilterStatePickerComponent } from "./state-picker-control";
+import { AdvancedFilterResultPickerComponent } from "./result-picker-control/result-picker-control";
+import { AdvancedFilterStatePickerComponent } from "./state-picker-control/state-picker-control";
 
 const components = [
     AdvancedFilterComponent,
