+++ src/@batch-flask/ui/advanced-filter/index.ts
@@ -2,6 +2,6 @@ export * from "./advanced-filter";
 export * from "./advanced-filter.component";
 export * from "./advanced-filter.module";
 export * from "./control-base";
-export * from "./state-picker-control";
-export * from "./result-picker-control";
+export * from "./state-picker-control/state-picker-control";
+export * from "./result-picker-control/result-picker-control";
 export * from "./list-filter-control/list-filter-control";
