+++ src/@batch-flask/ui/advanced-filter/result-picker-control/result-picker-control.scss
@@ -1,0 +1,10 @@
+@import "app/styles/variables";
+
+bl-adv-filter-resultpicker {
+    legend {
+        color: $primary-color;
+    }
+    .odata-checkbox {
+        margin-right: 10px;
+    }
+}
