+++ src/@batch-flask/ui/advanced-filter/result-picker-control/result-picker-control.ts
@@ -1,8 +1,10 @@
 import { Component, HostBinding, Input } from "@angular/core";
 import { FormControl, FormGroup } from "@angular/forms";
 import * as FilterBuilder from "@batch-flask/core/filter-builder";
-import { AdvancedFilter } from "./advanced-filter";
-import { AdvancedFilterControlBase } from "./control-base";
+import { AdvancedFilter } from "../advanced-filter";
+import { AdvancedFilterControlBase } from "../control-base";
+
+import "./result-picker-control.scss";
 
 export class ResultPickerControl extends AdvancedFilterControlBase {
     constructor(label: string, public results: string[]) {
