+++ src/@batch-flask/ui/advanced-filter/state-picker-control/state-picker-control.ts
@@ -1,8 +1,10 @@
 import { Component, HostBinding, Input } from "@angular/core";
 import { FormControl, FormGroup } from "@angular/forms";
 import * as FilterBuilder from "@batch-flask/core/filter-builder";
-import { AdvancedFilter } from "./advanced-filter";
-import { AdvancedFilterControlBase } from "./control-base";
+import { AdvancedFilter } from "../advanced-filter";
+import { AdvancedFilterControlBase } from "../control-base";
+
+import "./state-picker-control.scss";
 
 export class StatePickerControl extends AdvancedFilterControlBase {
     constructor(label: string, public states: string[]) {
