+++ src/app/styles/base/advanced-filter.scss
@@ -20,13 +20,3 @@
         }
     }
 }
-
-bl-adv-filter-statepicker, bl-adv-filter-resultpicker {
-    legend {
-        color: $primary-color;
-    }
-    .odata-checkbox {
-        margin-right: 10px;
-    }
-}
-
